+++ lib/screen_size_adapter.dart
@@ -1,15 +1,16 @@
-
 import 'dart:async';
 import 'dart:collection';
-import 'dart:io';
 import 'dart:ui';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 
+part 'src/config.dart';
 part 'src/screen_size_helper.dart';
 part 'src/screen_size_widget.dart';
 part 'src/screen_size_widget_flutter_binding.dart';
 part 'src/media_query_ext.dart';
 part 'src/dimension_ext.dart';
+part 'src/screen_size_adapter_controller.dart';